refactor: 여러 라인의 shared config import를 단일 라인으로 통합
- PATHS, TITLES 등 동일 모듈에서 가져오는 항목들을 하나의 import 문으로 통합
