+++ client/src/pages/game/CreateRoom.tsx
@@ -1,7 +1,6 @@
 import { useState } from 'react';
-import { PATHS } from '@/shared/config';
+import { PATHS, TITLES } from '@/shared/config';
 import { Title } from '@/shared/ui';
-import { TITLES } from '@/shared/config/titles';
 import { CommonBtn } from '@/shared/ui';
 
 export default function CreateRoom() {
